FIX: PDF file extension regular expression (#11)
Fixed the regexp to check the file extension. `[pdf]+$` would match anything ending with at least one of the three letters.

Use the `i` modifier on the regexp to ensure we're case insensitive.

Also use `test` rather than `match` since we only care about whether it matches and not the actual match data.